refactor(admin): colocate admin-artists search schema with the page
Moves the admin-artists URL search schema out of the shared
src/lib/searchSchemas.ts into src/pages/admin/ArtistsManagement/,
next to the components that use it, to keep the already-large route
file smaller and the schema closer to its only consumer.
